move static-model-policy-plugin into mini-chat
