+++ apps/hyperspot-server/Cargo.toml
@@ -18,7 +18,7 @@ static-tenants = ["dep:static-tr-plugin"]
 static-authn = ["dep:static-authn-plugin"]
 static-authz = ["dep:static-authz-plugin"]
 static-credstore = ["dep:static-credstore-plugin"]
-mini-chat = ["dep:mini-chat", "dep:static-mini-chat-model-policy-plugin"]
+mini-chat = ["dep:mini-chat"]
 otel = ["modkit/otel"]
 
 [dependencies]
@@ -61,7 +61,6 @@ clap = { workspace = true }
 
 # Optional mini-chat module
 mini-chat = { package = "cf-mini-chat", path = "../../modules/mini-chat/mini-chat", optional = true }
-static-mini-chat-model-policy-plugin = { package = "cf-static-mini-chat-model-policy-plugin", path = "../../modules/mini-chat/plugins/static-model-policy-plugin", optional = true }
 
 # Optional example module
 users-info = { path = "../../examples/modkit/users-info/users-info", optional = true }
