+++ modules/mini-chat/mini-chat/src/infra/plugins/mod.rs
@@ -1,2 +1,4 @@
 pub mod static_audit;
 pub use static_audit::StaticMiniChatAuditPlugin;
+pub mod static_model_policy;
+pub use static_model_policy::StaticMiniChatModelPolicyPlugin;
