+++ modules/mini-chat/mini-chat/src/infra/plugins/static_model_policy/config.rs
