+++ modules/mini-chat/mini-chat/src/infra/plugins/static_model_policy/mod.rs
@@ -1,7 +1,7 @@
-#![cfg_attr(coverage_nightly, feature(coverage_attribute))]
-
 pub mod config;
-pub mod domain;
 pub mod module;
+pub mod service;
+#[cfg(test)]
+mod tests;
 
 pub use module::StaticMiniChatModelPolicyPlugin;
