+++ modules/mini-chat/mini-chat/src/lib.rs
@@ -6,6 +6,7 @@ pub use module::MiniChatModule;
 
 // === PLUGIN MODULES ===
 pub use infra::plugins::StaticMiniChatAuditPlugin;
+pub use infra::plugins::StaticMiniChatModelPolicyPlugin;
 
 // === INTERNAL MODULES ===
 #[doc(hidden)]
