--- modules/mini-chat/plugins/static-model-policy-plugin/Cargo.toml
@@ -1,47 +1,0 @@
-[package]
-name = "cf-static-mini-chat-model-policy-plugin"
-version.workspace = true
-publish = false
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-description = "Static model-policy plugin for mini-chat with config-driven model catalog"
-
-[lib]
-name = "static_mini_chat_model_policy_plugin"
-
-[lints]
-workspace = true
-
-[dependencies]
-# Plugin SDK
-mini-chat-sdk = { package = "cf-mini-chat-sdk", version = "0.6.1", path = "../../mini-chat-sdk" }
-types-registry-sdk = { package = "cf-types-registry-sdk", version = "0.1.5", path = "../../../system/types-registry/types-registry-sdk" }
-
-# ModKit dependencies
-modkit = { workspace = true }
-modkit-macros = { workspace = true }
-
-# Async runtime
-async-trait = { workspace = true }
-tokio-util = { workspace = true }
-
-# Data structures
-uuid = { workspace = true }
-time = { workspace = true }
-
-# Error handling
-anyhow = { workspace = true }
-
-# Serialization
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-# Logging
-tracing = { workspace = true }
-
-# Required by modkit::module macro
-inventory = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true }
