--- modules/mini-chat/plugins/static-model-policy-plugin/src/domain/mod.rs
@@ -1,4 +1,0 @@
-mod client;
-pub mod service;
-
-pub use service::Service;
